refactor: move component ButtonDark.vue from TopBar to fixed position in bottom right of HomeView
- Removed component from the topbar
- Updated styles to position the component in the bottom right corner of HomeView
- Set the component to be fixed during scrolling for better accessibility
